fixing potentially dangerous bug

+++ meteor_setup.sh
@@ -1,3 +1,7 @@
+if [[ $# -eq 0 ]] ; then
+    echo 'Please enter a project name'
+    exit 0
+fi
 meteor create $1
 rm $1/$1.*
 mkdir $1/server $1/client $1/public $1/private
@@ -14,7 +18,7 @@ wget https://raw.githubusercontent.com/albertorios/meteor_tools/master/owners
 mv owners private/owners
 mv db_template server/db_template
 chmod 755 create_survey_db.sh
-meteor add iron:router semantic:ui flemay:less-autoprefixer fourseven:scss
+meteor add iron:router semantic:ui flemay:less-autoprefixer fourseven:scss accounts-password accounts-ui
 meteor remove autopublish insecure
 okgrow:router-autoscroll
 python setup.py
